move to sapi/cli
