--- conf.d/swoole_cli.php
@@ -1,9 +1,0 @@
-<?php
-
-use SwooleCli\Library;
-use SwooleCli\Preprocessor;
-use SwooleCli\Extension;
-
-return function (Preprocessor $p) {
-    $p->addExtension((new Extension('swoole_cli')));
-};
