+++ ext/phar/phar_internal.h
@@ -617,4 +617,4 @@ phar_path_check_result phar_path_check(char **p, size_t *len, const char **error
 
 END_EXTERN_C()
 
-#include "ext/swoole_cli/hook_stream.h"
+#include "sapi/cli/sfx/hook_stream.h"
