--- ext/swoole_cli/swoole_cli.stub.php
@@ -1,3 +1,0 @@
-<?php
-
-/** @generate-class-entries */
