--- ext/swoole_cli/swoole_cli_arginfo.h
@@ -1,4 +1,0 @@
-/* This is a generated file, edit the .stub.php file instead.
- * Stub hash: be039c47ec1997922d99134d3cd03579dd0fab8f */
-
-
