--- ext/swoole_cli/tests/001.phpt
@@ -1,10 +1,0 @@
---TEST--
-test
---EXTENSIONS--
-swoole_cli
---FILE--
-<?php
-echo "swoole-cli";
-?>
---EXPECTF--
-swoole-cli
