+++ sapi/cli/sfx/hook_cli.h
