+++ sapi/cli/sfx/hook_stream.h
