+++ sapi/cli/util.c
