+++ sapi/cli/util.h
